作品記録の評価の表示方法を修正
