+++ app/components/contents/record_content_component.rb
@@ -16,15 +16,23 @@ module Contents
           unless @record.instant?
             h.html(SpoilerGuardComponent.new(view_context, record: @record).render { |h|
               h.tag :div, class: "c-record-content__wrapper mb-3" do
-                if @record.rating
-                  h.html Labels::RatingLabelComponent.new(view_context, rating: @record.rating, advanced_rating: @record.advanced_rating, class_name: "mb-1").render
-                end
+                if @record.episode_record?
+                  if @record.rating
+                    h.html Labels::RatingLabelComponent.new(view_context, rating: @record.rating, advanced_rating: @record.advanced_rating, class_name: "mb-1").render
+                  end
 
-                if @record.work_record?
-                  h.html Collapses::DeprecatedRatingsCollapseComponent.new(view_context, record: @record, class_name: "ms-2 text-muted").render
-                end
+                  h.html BodyV6Component.new(view_context, content: @record.body, format: :markdown, height: 300).render
+                else
+                  h.tag :div, class: "row g-3" do
+                    h.tag :div, class: "col-12 col-md-4 col-xl-3 order-1 order-md-2" do
+                      h.html LabelGroups::WorkRecordRatingLabelGroupComponent.new(view_context, record: @record).render
+                    end
 
-                h.html BodyV6Component.new(view_context, content: @record.body, format: :markdown, height: 300).render
+                    h.tag :div, class: "col-12 col-md-8 col-xl-9 order-2 order-md-1" do
+                      h.html BodyV6Component.new(view_context, content: @record.body, format: :markdown, height: 300).render
+                    end
+                  end
+                end
               end
             })
           end
